+++ nflow-rest-api-common/src/main/java/io/nflow/rest/v1/msg/ListWorkflowInstanceResponse.java
@@ -28,7 +28,7 @@ public class ListWorkflowInstanceResponse extends ModelObject {
   public String type;
 
   @ApiModelProperty(value = "Workflow instance priority", required = true)
-  public Integer priority;
+  public Short priority;
 
   @ApiModelProperty("Parent workflow instance id for child workflows")
   public Long parentWorkflowId;
